changes back

+++ money-tracker-backend/src/main/java/com/moneyTracker/controllers/ProfileController.java
@@ -19,13 +19,13 @@ public class ProfileController {
     }
 
     @PutMapping("/{id}/balance")
-    public void updateProfileBalance(@PathVariable("id") final int profileId, @RequestParam("balance") final Long balance) {
+    public void updateProfileBalance(@PathVariable("id") final int profileId, @RequestParam("balance") final Double balance) {
         profileService.updateProfileBalance(profileId, balance);
     }
 
     @GetMapping("/{id}/balance")
-    public Long getProfileBalance(@PathVariable("id") final int id) {
+    public Double getProfileBalance(@PathVariable("id") final int id) {
         ProfileEntity profile = profileService.getProfileById(id);
-        return profile != null ? profile.getBalance() : 0L;
+        return profile != null ? profile.getBalance() : 0.0;
     }
 }
